+++ cpp08/ex01/span.cpp
@@ -6,7 +6,7 @@
 /*   By: gsharony <gsharony@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/12 14:10:59 by gsharony          #+#    #+#             */
-/*   Updated: 2020/12/16 15:27:52 by gsharony         ###   ########.fr       */
+/*   Updated: 2020/12/16 15:59:32 by gsharony         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,15 +49,15 @@ void			Span::addNumber(int n)
 	this->_array.push_back(n);
 }
 
-int				Span::shortestSpan(void)
+long			Span::shortestSpan(void)
 {
 	if (this->_array.size() <= 1)
 		throw Span::ArrayIsTooShort();
-	int		t;
-	int		r;
+	long	t;
+	long	r;
 	for (size_t i = 0; i < this->_array.size() - 1; i++)
 	{
-		t = this->_array[i + 1] - this->_array[i];
+		t = static_cast<long>(this->_array[i + 1]) - static_cast<long>(this->_array[i]);
 		t = (t < 0) ? -t : t;
 		if (!i || (i && t < r))
 			r = t;
@@ -65,13 +65,13 @@ int				Span::shortestSpan(void)
 	return (r);
 }
 
-int				Span::longestSpan(void)
+long			Span::longestSpan(void)
 {
 	if (this->_array.size() <= 1)
 		throw Span::ArrayIsTooShort();
 	std::vector<int>::iterator min = std::min_element(this->_array.begin(), this->_array.end());
 	std::vector<int>::iterator max = std::max_element(this->_array.begin(), this->_array.end());
-	return (*max - *min);
+	return (static_cast<long>(*max) - static_cast<long>(*min));
 }
 
 const char		*Span::ArrayIsFull::what() const throw()
